requirements.txt
Django==1.10
django-ordered-model==1.2.1
gunicorn==19.6.0
psycopg2==2.6.2
